Refactor profile comp
